Add progress dialog resource
AP Suite has convenience tools for displaying a modal progress
dialog to wrap asyncronous or long-running actions.

While this functionality could be replicated using native WPF controls
and the Prisim provided dialog service, it can be difficult to make robust.

Providing the existing functionality as a resource allows for easier
development of a custom analysis that includes long-running work,
and additionally ensures the displayed dialog will appear with
native theme and style for integration with the main application.