Update and add events
Refactor events to follow subject/verb patterns for clarity.
Supported subjects are analysis tree, node, and view model.
Analysis set is renamed to analysis tree to match application.
Use delete over close where applicable to indicate that the resource
is unrecoverable. For some subjects such as view models, this is
effectively a close, but the instance is intended to be transitory so
explicitly calling a delete is a more accurate representation.

Events are either command+subject to request and action or
subject+past tense verb to hook into an action that occured.

Removed distinction between create and load. Consolidate under create
as an instance is always being created, but add a event trigger property
to identify the source of creation for optional routing.

Added one pre-action event hook for analysis tree closing, allowing a
hook to retrieve data if necessary. By the time the closed event is
published, relevant data may be disposed. The addition of closing
is useful for performing actions such as an export of conditional
cleanup based on data or state before analysis tree close if finalized.

Events that may not yet have full support may be included, especially
regarding analysis tree events.