Fix histogram data interface
BinCenter only makes sense as a method accepting a bin index.
However it is trivially computed from a start value and the bin
width as the histogram is fixed width.

TotalCount is a contextual optimization in AP Suite and should not
be exposed via extensions.

BinCount is trivially give from Values.Length, and better to remove
unnecessary duplication