Add DataSectionName used for nodes
Each analysis node instance in AP Suite will be given a managed unique
section name that can be used to write data to the underlying file. This
allows a mechanism by which AP Suite can consistently identify sections
created by extensions, and this section can have stronger guarantees of
consistency by handled by the application instead of requiring
extensions to have a more full understanding of the lifecycle. This will
get mapped to APT Section Header Section Type field, so it should be no
longer than 32 characters.

Name is not a unique default implementation, but there is no better
alternative. I think it might be good enough in practice, and prefer
good enough to throwing the exception.