Add extension options feature
Extensions can register an options model to integration with AP Suite
options infrastructure, allowing editing and persisting global options
for an extension in a standard and expected place in the application.

Options are hosted under Configuration > Options in AP Suite like
other native AP Suite options.

Options can be registered using only a model, in which case a default
editor is generated, and public properties of the options class can be
customized with attributes to control the editor used for the basic data
type. This currently support file and folder paths for string
properties, both single or multiple selections supported. If multiple
selections, the resulting string is delimited with Path.PathSeparator,
by default a semicolon.

Options can be accessed from the hosted analysis in the extension by
injecting and using the IOptionsAccessor interface and requesting
the registered options model.
the IOptionsAccessor ser