Merge Cameca.Extensions.Control
Conceptually this project served the same purpose as this interface:
map internal AP Suite objects into fixed public interfaces.
Originally this project was split as this library was intented strictly
for custom analyses, not for other potential future interfaces such
as reconstructions, while the controls could potentially be used for
essentially any kind of AP Suite extension.

In practice this just increases complexity and causes issues when
interfaces really ought to be present in both. Bringing the controls
into this library practically makes this the de facto dependency for
all AP Suite extensions, not just custom analyses. While the name
will remain CustomAnalysis.Interface as changing that would introduce
challanges for upgrading, and the goal is to minimize complexity,
moving forward this dependency should include all the code for
direct interfacing between AP Suite and any extension.