Add value bounds to color map
Color maps have ambiguous mappings without the bounded mapped values.
This wasn't included initially as in AP Suite as the bounds are wrapped
in a container class, but for the purpose of the extensions this should
always be included so it is set directly in the color map.

Set defaults of 0-100 if not value bound are explicitly set in an
extension. This matches the AP Suite defaults.